# fix(bench): explicit timers don't force high precision measurements