--- a/cli/tests/integration/bench_tests.rs
+++ b/cli/tests/integration/bench_tests.rs
@@ -187,6 +187,11 @@ itest!(bench_explicit_start_end {
   exit_code: 1,
 });
 
+itest!(bench_explicit_start_end_low_precision {
+  args: "bench --quiet -A bench/explicit_start_and_end_low_precision.ts",
+  output: "bench/explicit_start_and_end_low_precision.out",
+});
+
 itest!(bench_with_config {
   args: "bench --config bench/collect/deno.jsonc bench/collect",
   exit_code: 0,