--- a/cli/tests/testdata/bench/explicit_start_and_end_low_precision.out
+++ b/cli/tests/testdata/bench/explicit_start_and_end_low_precision.out
@@ -5,5 +5,5 @@ runtime: deno [WILDCARD] ([WILDCARD])
 benchmark                    time (avg)        iter/s             (min … max)       p75       p99      p995
 ----------------------------------------------------------------------------- -----------------------------
 noop with start and end  [WILDCARD] [WILDCARD]/iter[WILDCARD]([WILDCARD] … [WILDCARD]) [WILDCARD]
-Warning start() and end() calls in "noop with start and end" are ignored because it averages less
+Warning: start() and end() calls in "noop with start and end" are ignored because it averages less
 than 0.01s per iteration. Remove them for better results.