--- a/cli/tests/testdata/bench/explicit_start_and_end_low_precision.ts
+++ b/cli/tests/testdata/bench/explicit_start_and_end_low_precision.ts
@@ -1,0 +1,4 @@
+Deno.bench("noop with start and end", (b) => {
+  b.start();
+  b.end();
+});