@bartlomieju 2023-08-27
@nayeemrmn I'm afraid this test is flaky, eg: https://github.com/denoland/deno/actions/runs/5990043261/job/16246927846?pr=18642 from https://github.com/denoland/deno/pull/18642